requirements.txt
basicauth==0.2
certifi==2015.9.6.2
cloudinary==1.1.3
coverage==3.7.1
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-allauth==0.23.0
django-crispy-forms==1.5.0
django-haystack==2.4.0
django-inplaceedit==1.4.1
django-markdown==0.8.4
django-nose==1.4.1
django-registration-redux==1.2
django-smtp-ssl==1.0
django-tables2==1.0.4
django-toolbelt==0.0.1
elasticsearch==1.6.0
gunicorn==19.3.0
httplib2==0.9.1
Markdown==2.6.2
mysql-connector-python==2.0.4
nose==1.3.7
oauth==1.0.1
oauthlib==1.0.3
pillow==2.9.0
psycopg2==2.6.1
pyelasticsearch==1.4
PyJWT==1.4.0
python-gettext==2.1
python-oauth2==0.7.0
python-openid==2.2.5
python-social-auth==0.2.12
python3-openid==3.0.6
requests==2.7.0
requests-oauthlib==0.5.0
simplejson==3.8.0
six==1.9.0
static3==0.6.1
urllib3==1.12
virtualenv==13.1.0
